requirements/local.txt
-r base.txt #includes the base.txt requirements file

django-debug-toolbar==1.0.1
django-extensions==1.3.3
Werkzeug==0.9.4